refactor methods in UserController

+++ src/main/java/com/example/emag/model/dao/IAddressDAO.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 public interface IAddressDAO {
 
-    void addAddress(Integer userId, Address address) throws SQLException;
-    List<Address> getAllAddresses(Integer userId) throws SQLException;
-    void updateAddress(Address address, Integer addressId) throws SQLException;
-    Address getAddress(Integer addressId) throws SQLException;   //not necessary???
-    void deleteAddress(Integer addressId) throws SQLException;
+    void addAddress(long userId, Address address) throws SQLException;
+    List<Address> getAllAddresses(long userId) throws SQLException;
+    void updateAddress(Address address, long addressId) throws SQLException;
+    Address getAddress(long addressId) throws SQLException;   //not necessary???
+    void deleteAddress(long addressId) throws SQLException;
 }
